requirements.txt
appnope==0.1.2
asgiref==3.4.1
backcall==0.2.0
Currency==0.1.2
decorator==5.0.9
defusedxml==0.7.1
diff-match-patch==20200713
Django==3.2.5
django-admin-rangefilter==0.8.1
django-debug-toolbar==3.2.2
django-extensions==3.1.3
django-import-export==2.6.0
et-xmlfile==1.1.0
flake8==3.9.2
flake8-builtins==1.5.3
flake8-import-order==0.18.1
flake8-print==4.0.0
ipython==7.26.0
ipython-genutils==0.2.0
jedi==0.18.0
MarkupPy==1.14
matplotlib-inline==0.1.2
mccabe==0.6.1
odfpy==1.4.1
openpyxl==3.0.9
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.20
ptyprocess==0.7.0
pycodestyle==2.7.0
pyflakes==2.3.1
Pygments==2.10.0
pytz==2021.1
PyYAML==5.4.1
six==1.16.0
sqlparse==0.4.1
tablib==3.0.0
traitlets==5.0.5
views==0.3
wcwidth==0.2.5
xlrd==2.0.1
xlwt==1.3.0
